Title:chore check in 5.2.0.36 version
Description:

RootCause:Null

Solution:
Reviewed-on: http://gerrit.uniontech.com/c/chengdu/dde-file-manager/+/3707
Tested-by: <mailman@uniontech.com>

+++ dde-desktop/dbus/filedialog/dbusfiledialoghandle.cpp
@@ -145,7 +145,7 @@ void DBusFileDialogHandle::setOption(int option, bool on)
 
 int DBusFileDialogHandle::options() const
 {
-    return DFileDialogHandle::options();
+    return static_cast<int>(DFileDialogHandle::options());
 }
 
 bool DBusFileDialogHandle::testOption(int option) const
